Refactor and introduce class ActorEvents
ActorEvents can be used to retrieve events of a ModelActor from the engine's event journal.
Earlier classes CaseHierarchy, ProcessIdentifier and ActorIdentifier were actually mostly reading events from the event journal of the engine.
This is now much more explicit in the class name, and also the classes have their own files now.
Furthermore the same is introduce for Tenants and ConsentGroups.