Introduce AsyncError (async stacktrace printing)
Added slightly better stack trace printing when the framework encounters assertion errors.
On async calls it is hard to find out the actual stack trace.
caseservice.ts now uses the new trick to print the location that triggered a failing assertion.
Rest of async calls of the framework are yet to start using the same trick.