Loading sub case events should not assert state
When a sub case is loaded, the state of the CaseTask in the parent may differ from the state of the CasePlan in the subcase.
E.g. because the child is earlier in a different state than the parent is.
Until now when loading case events the algoritm presumed both have exact same state.
This condition is removed, as with the new fault bubbling, the chance of difference is more likely than before.