JSON to type parsing now handles null
In the old API, always a new object was created and copied to (even when a null object was given back from the server).
WIth the new typing, this no longer happens, so it failed on APIs that return a null object.
This commit fixes that

Also cleaned imports in TestTaskCountAPI