Make TestMoveTaskMigration little more robust
The name of the case plan in the model is "migration/move_task_v0". Due to an issue in Path serializing, the name is interpreted as "move_task_v0", but that is not the real name. With upcoming cluster implementation this is corrected, and then this test will fail. Making it more robust by reading the actual case plan